Many minor changes / corrections.
One major change:
This is now a complete Eclipse project (Eclipse Kepler with CDT), so you
can either import it into Eclipse or you may just use the make files in
the RELEASE or DEBUG directories without Eclipse.

Tested on a Debian 7 system. Compiles with GCC 4.7.2. Does not compile
on Debian 6 GCC 4.4.5 any more due to changes in C++11 support of GCC.